drm: rp1: rp1-dpi: Experimentally add a new RGB666 pin-out variant
The new format resembles "mode 6" but moves R high:
   R: GPIOs 27:22 (gap 21:18)
   G: GPIOs 17:12 (gap 11:10)
   B: GPIOs  9: 4

No MEDIA_BUS_FMT_* code is yet defined for it, but it can be tested
using any unsupported format such as e.g. MEDIA_BUS_FMT_GBR888_1X24.

Also define some pinctrls for it.

This has not been tested and will not be maintained or supported.
There are NO plans to merge this into a released kernel.