drm: rp1: rp1-dpi: First rough / PoC version of interlaced DPI output
(Since the "simple" DRM pipeline seems not to support interlace, and
panel timings have no flag for CSync, auto-infer these options when
horizontal rate is ~15kHz and line count is either 525 or 625.)

Bodge the DPI hardware driver to output alternate fields, exploiting
its ability to change VFP while running. However, it still cannot
generate a correct VSync for interlace.

Instead, we replace VSync and HSync by "helper" signals. The idea is
to use RP1 PIO to snoop on GPIOs 2,3 and re-create CSync on GPIO 1.
The PIO program is *not* currently integrated with the DPI driver
and needs to be invoked from user-land.

Experimentally, add a nonstandard RGB666 pinout resembling "mode 6"
but with the Red bits higher, leaving the PCM/I2S bus free.